src/requirements.txt
flask==2.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.1
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
python-decouple==3.5
gunicorn==20.1.0
jinja2==3.0.3
flask-restx==0.5.1
flask_bcrypt==0.7.1
Werkzeug==2.0.3
pandas==1.4.3
boto3==1.24.33
botocore==1.27.33
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-billing==1.6.1
google-cloud-core==2.3.1
google-cloud-datastore==2.7.2
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
jellyfish==0.9.0
jmespath==1.0.1
proto-plus==1.20.6
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.10